requirements.txt
streamlit
Pillow
numpy
tensorflow
protobuf==3.20.0

